fix:move component to folder
