+++ client/src/App.js
@@ -3,7 +3,6 @@ import "./App.css";
 import { CircularProgress, CircularProgressLabel } from "@chakra-ui/react";
 import { Button, ButtonGroup } from "@chakra-ui/react";
 import { Text } from "@chakra-ui/react";
-import SliderImage from "./components/SliderImage";
 import UnauthenticatedApp from "./pages/UnauthenticatedApp";
 import { useAuth } from "./contexts/authentication";
 import AuthenticatedApp from "./pages/AuthenticatedApp";
