+++ client/src/components/Homepage/CustomerCarousel.js
