+++ client/src/components/Homepage/SliderImage.js
